backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
pymupdf
python-docx
python-multipart
google-genai
python-jose[cryptography]
